[Fix] Fix bugs in RLE training (#1584)
* fix bugs for training

* modify configs

* pass unittest

* shift in tta as default

* fix sigmoid order

+++ mmpose/codecs/regression_label.py
@@ -51,7 +51,7 @@ class RegressionLabel(BaseKeypointCodec):
                  (keypoints <= [w - 1, h - 1])).all(axis=-1) & (
                      keypoints_visible > 0.5)
 
-        reg_labels = keypoints / np.array([w, h])
+        reg_labels = (keypoints / np.array([w, h])).astype(np.float32)
         keypoint_weights = np.where(valid, 1., 0.).astype(np.float32)
 
         return reg_labels, keypoint_weights
